Fix store images not loading when running manage.py from outside project path
Before this fix, when an image was loaded, the path for the image was calculated
as directory_path+'/'+image_name, which meant that it ended up being a relative
path resolved against the directory in which manage.py was executed. That is, if
one executed manage.py in a directory other than the project root, it would
attempt to load the image from the directory in which it was executed.

Now, it is calculated directly using os.path.join(PROJECT_PATH, directory,
image_name), so it always is resolved against the project root, as expected.