Made companion fabric images be stored in media/companion/fabrics
Previously, they were one of the many things stored in avatars/. I also
removed the getImage view from the companion app's views.py, since it
should (in theory?) no longer be necessary.

I was not able to test this commit because of the currently-unfinished
state of the companion app. However, I followed the same general
principles as the commit that moved the store product images, so it
should be OK.