Removed database from Git
Now, there is a shell script updatedb.sh in the project root. When you
execute this script (via: ./updatedb.sh), Django will re-create the
database from the models found in the application directories (login,
companion, blog, webstore, frontend). No other tables will be changed.

Note: if you add another 'app' to the site and you want its DB to be
affected by the updatedb script, you will need to add it to the array of
apps in the updatedb.sh file.

Note 2: There is also an updatedb.bat for Windows users ...